--- a/src/equation_of_state.hpp
+++ b/src/equation_of_state.hpp
@@ -334,7 +334,7 @@ class DryAir : public GasMixture {
                                                                        const double &Temp, const double &gamma,
                                                                        const double &Rg, const int &num_equation,
                                                                        const int &dim) {
-    for (int d = 0; d < dim; d++) stagState[1+d] = 0.;
+    for (int d = 0; d < dim; d++) stagState[1 + d] = 0.;
     stagState[1 + dim] = Rg / (gamma - 1.) * stateIn[0] * Temp;
   }
 
@@ -354,9 +354,10 @@ class DryAir : public GasMixture {
     if (thrd == 0) stateOut[1 + dim] = p / (gamma - 1.) + ke;
   }
 
-  static MFEM_HOST_DEVICE void modifyEnergyForPressure_gpu_serial(const double *stateIn, double *stateOut, const double &p,
-                                                                  const double &gamma, const double &Rg,
-                                                                  const int &num_equation, const int &dim) {
+  static MFEM_HOST_DEVICE void modifyEnergyForPressure_gpu_serial(const double *stateIn, double *stateOut,
+                                                                  const double &p, const double &gamma,
+                                                                  const double &Rg, const int &num_equation,
+                                                                  const int &dim) {
     double ke;
     ke = 0.;
     for (int d = 0; d < dim; d++) ke += stateIn[1 + d] * stateIn[1 + d];