--- a/src/equation_of_state.hpp
+++ b/src/equation_of_state.hpp
@@ -353,6 +353,19 @@ class DryAir : public GasMixture {
 
     if (thrd == 0) stateOut[1 + dim] = p / (gamma - 1.) + ke;
   }
+
+  static MFEM_HOST_DEVICE void modifyEnergyForPressure_gpu_serial(const double *stateIn, double *stateOut, const double &p,
+                                                                  const double &gamma, const double &Rg,
+                                                                  const int &num_equation, const int &dim) {
+    double ke;
+    ke = 0.;
+    for (int d = 0; d < dim; d++) ke += stateIn[1 + d] * stateIn[1 + d];
+    ke *= 0.5 / stateIn[0];
+
+    for (int eq = 0; eq < num_equation; eq++) stateOut[eq] = stateIn[eq];
+
+    stateOut[1 + dim] = p / (gamma - 1.) + ke;
+  }
 #endif
 };