@trevilo 2022-05-11
The test failures in `perfect_gas.test` were due to 

1. randomness in the test conditions occasionally leading to loss of precision, which is orthogonal to this PR and which @dreamer2368 is addressing in PR #129.
2. `die.test`... which is always a bit finnicky and which @koomie and I discussed deactivating to avoid false failures.

So... this is ready to review and merge from my perspective.