@koomie 2022-05-12
> The test failures were in
> 
> 1. `perfect_gas.test`, due to randomness in the test conditions occasionally leading to loss of precision, which is orthogonal to this PR and which @dreamer2368 is addressing in PR [perfect_gas.test - Fixes to decrease false failure rates. #129](https://github.com/pecos/tps/pull/129).
> 2. `die.test`... which is always a bit finnicky and which @koomie and I discussed deactivating to avoid false failures.
> 
die.test just disabled in main branch.